Many improvements and bug fixes

+++ ChiefsPlayer/Source/CVideoView Progress Bar Delegate.swift
@@ -47,7 +47,7 @@ extension CVideoView:CVideoProgressViewDelegate {
         }
         var duration = CGFloat(playerDuration)
         if duration.isNaN || duration.isInfinite {duration = 0} //Happens when user is on error only
-        let selectedSeconds = Int(duration * percent)
+        let selectedSeconds = Double(duration * percent)
         let selectedTime = CMTime(seconds: Double(selectedSeconds), preferredTimescale: 1)
         ChiefsPlayer.shared.player.seek(to: selectedTime)
     }
